removed 'writer' related code as we only do formatting within the eventlogger
